requirements.txt
streamlit>=1.20.0
plotly>=5.0.0
pandas>=1.3.0
numpy>=1.21.0
cassandra-driver>=3.25.0
kafka-python>=2.0.0
pyspark>=3.2.0
requests>=2.25.0
